+++ NEWS.md
@@ -1,4 +1,4 @@
-# CUSOMShinyHelpers 1.6.1
+# CUSOMShinyHelpers 1.6.3
 
  
 ## New Functionality